+++ facade/src/main/scala/react/sortable/SortableContainer.scala
@@ -113,12 +113,14 @@ object SortableContainer {
         .asInstanceOf[Props]
   }
 
-  /**
-    * Wrap another component
+  /** Wrap another component
     *
-    * @param wrappedComponent The wrapped component itself
-    * @tparam P The type of Props of the wrapped component
-    * @return A component wrapping the wrapped component...
+    * @param wrappedComponent
+    *   The wrapped component itself
+    * @tparam P
+    *   The type of Props of the wrapped component
+    * @return
+    *   A component wrapping the wrapped component...
     */
   def wrap[P, CT[_, _]](
     wrappedComponent: GenericComponent[P, CT, _],