+++ facade/src/main/scala/react/sortable/SortableView.scala
@@ -15,7 +15,7 @@ object SortableView {
         ^.className := "sortable-handle",
         svg(
           ^.className := "sortable-handle-svg",
-          viewBox := "0 0 24 24",
+          viewBox     := "0 0 24 24",
           path(
             d := "M9,8c1.1,0,2-0.9,2-2s-0.9-2-2-2S7,4.9,7,6S7.9,8,9,8z M9,10c-1.1,0-2,0.9-2,2s0.9,2,2,2s2-0.9,2-2S10.1,10,9,10z M9,16c-1.1,0-2,0.9-2,2s0.9,2,2,2s2-0.9,2-2S10.1,16,9,16z"
           ),