+++ project/Settings.scala
@@ -1,0 +1,12 @@
+object Settings {
+ object LibraryVersions {
+   val reactJS          = "17.0.2"
+   val scalaJsReact     = "2.0.0-RC3"
+   val reactSortableHOC = "2.0.0"
+   val scalaJSDom       = "1.2.0"
+   val utest            = "0.7.10"
+   val cats             = "2.6.1"
+   val reactVirtualized = "0.12.0"
+   val reactDraggable   = "0.13.0"
+ }
+}